Moved ServerHandler struct to be with the ServeHTTP method that it implements for code clarity.
